refactor: Change name of "Topic" component to "Feed"
